+++ source/common/http/header_utility.cc
@@ -14,15 +14,15 @@ namespace Http {
 //   exact string matching.
 //   This is now deprecated.
 // 2.header_match_specifier which can be any one of exact_match, regex_match or range_match.
+//   Each of these also can be inverted with the invert_match option.
 //   Absence of these options implies empty header value match based on header presence.
 //   a.exact_match: value will be used for exact string matching.
 //   b.regex_match: Match will succeed if header value matches the value specified here.
 //   c.range_match: Match will succeed if header value lies within the range specified
 //     here, using half open interval semantics [start,end).
-//   d.inverse_match: Match will succeed if the header is not present or the value is different than
-//     the one given.
 HeaderUtility::HeaderData::HeaderData(const envoy::api::v2::route::HeaderMatcher& config)
-    : name_(config.name()) {
+    : name_(config.name()),
+      invert_match_(PROTOBUF_GET_WRAPPED_OR_DEFAULT(config, invert_match, false)) {
   switch (config.header_match_specifier_case()) {
   case envoy::api::v2::route::HeaderMatcher::kExactMatch:
     header_match_type_ = HeaderMatchType::Value;
@@ -36,10 +36,6 @@ HeaderUtility::HeaderData::HeaderData(const envoy::api::v2::route::HeaderMatcher
     header_match_type_ = HeaderMatchType::Range;
     range_.set_start(config.range_match().start());
     range_.set_end(config.range_match().end());
-    break;
-  case envoy::api::v2::route::HeaderMatcher::kInverseMatch:
-    header_match_type_ = HeaderMatchType::Inverse;
-    value_ = config.inverse_match();
     break;
   case envoy::api::v2::route::HeaderMatcher::HEADER_MATCH_SPECIFIER_NOT_SET:
     FALLTHRU;
@@ -90,13 +86,11 @@ bool HeaderUtility::matchHeaders(const Http::HeaderMap& request_headers,
         break;
       }
 
-      case HeaderMatchType::Inverse:
-        matches &= (header == nullptr) || (header->value() != cfg_header_data.value_.c_str());
-        break;
-
       default:
         NOT_REACHED;
       }
+
+      matches ^= cfg_header_data.invert_match_;
 
       if (!matches) {
         break;